requirements.txt
Django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
Pygments==2.16.1
httpie==3.2.2